:recycle: define mixin tests for Template and CRBanim readers
