+++ django-data/image/excel/helpers/__init__.py
@@ -6,9 +6,10 @@ Created on Fri Jul  5 16:30:14 2019
 @author: Paolo Cozzi <cozzi@ibba.cnr.it>
 """
 
-from .exceltemplate import ExcelTemplate, TEMPLATE_COLUMNS
+from .exceltemplate import ExcelTemplateReader, TEMPLATE_COLUMNS
 from .exceptions import ExcelImportError
 from .fill_uid import upload_template
 
 __all__ = [
-    "ExcelTemplate", "TEMPLATE_COLUMNS", "ExcelImportError", "upload_template"]
+    "ExcelTemplateReader", "TEMPLATE_COLUMNS", "ExcelImportError",
+    "upload_template"]
