+++ django-data/image/image_app/tests/__init__.py
@@ -6,6 +6,6 @@ Created on Wed Feb 14 10:33:34 2018
 @author: Paolo Cozzi <paolo.cozzi@ptp.it>
 """
 
-from .mixins import DataSourceMixinTestCase
+from .mixins import DataSourceMixinTestCase, FileReaderMixinTestCase
 
-__all__ = ["DataSourceMixinTestCase"]
+__all__ = ["DataSourceMixinTestCase", "FileReaderMixinTestCase"]
