+++ django-data/image/image_app/tests/mixins.py
@@ -8,7 +8,8 @@ Created on Fri Jul  5 15:57:10 2019
 
 from common.tests import DataSourceMixinTestCase as CommonDataSourceMixinTest
 
-from ..models import db_has_data, Animal, Sample, Submission
+from ..models import (
+    db_has_data, Animal, Sample, Submission, DictCountry, DictSpecie, DictSex)
 
 
 class DataSourceMixinTestCase(CommonDataSourceMixinTest):
@@ -34,3 +35,45 @@ class DataSourceMixinTestCase(CommonDataSourceMixinTest):
         self.assertFalse(db_has_data())
         self.assertFalse(Animal.objects.exists())
         self.assertFalse(Sample.objects.exists())
+
+
+class FileReaderMixinTestCase:
+    """A class to deal with common text between CRBAnimReader and
+    ExcelTemplateReader"""
+
+    def test_check_species(self):
+        """Test check species method"""
+
+        # get a country
+        country = DictCountry.objects.get(label="United Kingdom")
+
+        check, not_found = self.reader.check_species(country)
+
+        self.assertTrue(check)
+        self.assertEqual(len(not_found), 0)
+
+        # changing species set
+        DictSpecie.objects.filter(label='Capra hircus').delete()
+
+        check, not_found = self.reader.check_species(country)
+
+        # the read species are not included in fixtures
+        self.assertFalse(check)
+        self.assertGreater(len(not_found), 0)
+
+    def test_check_sex(self):
+        """Test check sex method"""
+
+        check, not_found = self.reader.check_sex()
+
+        self.assertTrue(check)
+        self.assertEqual(len(not_found), 0)
+
+        # changing sex set
+        DictSex.objects.filter(label='female').delete()
+
+        check, not_found = self.reader.check_sex()
+
+        # the read species are not included in fixtures
+        self.assertFalse(check)
+        self.assertGreater(len(not_found), 0)
